Removed Function.php
Removed Function.php
Added php files for all pages.
