+++ MainMenu.php
@@ -1,5 +1,5 @@
 <?php
-  require("Function.php");
+  require("MainMenuPHP.php");
   if (session_status() == PHP_SESSION_NONE) 
   {
     session_start();
@@ -17,7 +17,6 @@
 <!doctype html>
 <html lang="en">
   <head>
-    <!-- Required meta tags -->
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
